Migrator/salesforce query single (#429)
* Light refactoring

* Refactoring Create operation

* Refactoring Update and Upsert operation

* Refactoring Retrieve operation

* Refactoring Query operation

* Fix query tests -now the headers are migrated first-

* Replace renamed class

* refactored the way we build the transform message

* Resolve accessTokenId attribute inside the abstract class

* Fix Create operation description

* Added migrator for Query Single operation + update salesforce version
