+++ mule-migration-tool-library/src/main/java/com/mulesoft/tools/migration/library/mule/steps/salesforce/SalesforcePomContribution.java
@@ -28,7 +28,7 @@ public class SalesforcePomContribution implements PomContribution {
   @Override
   public void execute(PomModel object, MigrationReport report) throws RuntimeException {
     object.addDependency(new DependencyBuilder()
-        .withGroupId("com.mulesoft.connector")
+        .withGroupId("com.mulesoft.connectors")
         .withArtifactId("mule-salesforce-connector")
         .withVersion(targetVersion("mule-salesforce-connector"))
         .withClassifier("mule-plugin")
