+++ mule-migration-tool-library/src/main/java/com/mulesoft/tools/migration/library/mule/tasks/SalesforceMigrationTask.java
@@ -9,6 +9,7 @@ import com.mulesoft.tools.migration.library.mule.steps.salesforce.CachedBasicCon
 import com.mulesoft.tools.migration.library.mule.steps.salesforce.CreateJobOperation;
 import com.mulesoft.tools.migration.library.mule.steps.salesforce.CreateOperation;
 import com.mulesoft.tools.migration.library.mule.steps.salesforce.QueryOperation;
+import com.mulesoft.tools.migration.library.mule.steps.salesforce.QuerySingleOperation;
 import com.mulesoft.tools.migration.library.mule.steps.salesforce.RetrieveOperation;
 import com.mulesoft.tools.migration.library.mule.steps.salesforce.SalesforcePomContribution;
 import com.mulesoft.tools.migration.library.mule.steps.salesforce.UpdateOperation;
@@ -55,6 +56,7 @@ public class SalesforceMigrationTask extends AbstractMigrationTask {
                         new UpsertBulkOperation(),
                         new RetrieveOperation(),
                         new QueryOperation(),
+                        new QuerySingleOperation(),
                         new CachedBasicConfiguration(),
                         new SalesforcePomContribution(),
                         new CreateJobOperation());
