+++ mule-migration-tool-library/src/test/resources/mule/apps/salesforce/salesforce-querySingleNative-original.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<mule xmlns:sfdc="http://www.mulesoft.org/schema/mule/sfdc" xmlns="http://www.mulesoft.org/schema/mule/core" xmlns:doc="http://www.mulesoft.org/schema/mule/documentation"
+      xmlns:spring="http://www.springframework.org/schema/beans"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-current.xsd
+http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd
+http://www.mulesoft.org/schema/mule/sfdc http://www.mulesoft.org/schema/mule/sfdc/current/mule-sfdc.xsd">
+
+    <flow name="salesforce-querySingleNative-flow">
+        <sfdc:query-single config-ref="Salesforce__Basic_Authentication"  doc:name="Salesforce" doc:description="Notes" query="SELECT Id,Name FROM Account" >
+            <sfdc:headers ref="#[flowVars.headers]"/>
+        </sfdc:query-single>
+    </flow>
+
+</mule>
