+++ mule-migration-tool-library/src/test/resources/mule/apps/salesforce/salesforce-querySingleWithAccessTokenId.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<mule xmlns="http://www.mulesoft.org/schema/mule/core" xmlns:doc="http://www.mulesoft.org/schema/mule/documentation" xmlns:salesforce="http://www.mulesoft.org/schema/mule/salesforce" xmlns:sfdc="http://www.mulesoft.org/schema/mule/sfdc" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.mulesoft.org/schema/mule/sfdc http://www.mulesoft.org/schema/mule/sfdc/current/mule-sfdc.xsd http://www.mulesoft.org/schema/mule/salesforce http://www.mulesoft.org/schema/mule/salesforce http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd">
+
+    <flow name="salesforce-querySingleWithAccessTokenId-flow">
+        <salesforce:query doc:name="Salesforce" doc:description="Notes" config-ref="Salesforce__Basic_Authentication" >
+            <!--Migration INFO: The access token id parameter was removed in salesforce 10.x-->
+            <!--<sfdc:query xmlns:sfdc="http://www.mulesoft.org/schema/mule/sfdc" xmlns:doc="http://www.mulesoft.org/schema/mule/documentation" config-ref="Salesforce__Basic_Authentication" doc:name="Salesforce" doc:description="Notes" query="SELECT Id,Name FROM Account" accessTokenId="accessTokenId" />-->
+            <salesforce:salesforce-query><![CDATA[SELECT Id,Name FROM Account LIMIT 1]]></salesforce:salesforce-query>
+        </salesforce:query>
+
+    </flow>
+
+</mule>
