+++ target-modules/src/main/resources/target-versions.properties
@@ -11,7 +11,7 @@ mule-sftp-connector=1.3.7
 mule-vm-connector=${mule.vm.connector.version}
 mule-objectstore-connector=1.1.5
 mule-aggregators-module=${mule.aggregators.module.version}
-mule-salesforce-connector=10.2.0
+mule-salesforce-connector=10.3.0
 
 mule-validation-module=${mule.validation.module.version}
 mule-scripting-module=${mule.scripting.module.version}
